Rename SiEffectorModule and SiOPMModule for clarity
SiEffectorModule is now SiEffector, all methods
accepting or returning effects were renamed as
well to avoid the word "effector".

SiOPMModule is now SiOPMSoundChip, which
is more appropriate. The entire /processor folder
is now /chip too.

These renames are necessary to better convey
meaning of each part of the system, and it's
important to do them before those classes
become public.