CI: Do string comparisons on boolean action inputs
Because booleans are a lie. Values passed externally
seem to be correctly typed, but defaults are converted
to strings. So checking a boolean input like a boolean
with default 'false' is useless, it will always be truthy.

Also action inputs don't have a type property, unlike
workflow inputs. God knows why...