Change chip type from String to a enum, rename global constants
This is a breaking, but necessary change. Stringy
chip type has been awkward from the beginning,
but I rolled with it for the time being. This design
suggests some flexibility for custom code, but
we don't really consider extensibility at this time.

As we're exposing new enums, it becomes clear
that some existing ones need to be renamed for
consistency and clarity. Numerical values stay
the same.

And finally, since we're improving on the voice
type, some additional methods have been exposed because they are useful when
defining voices.