Improve memory management of SiONData and related classes
Previously this code would have a cyclic reference where
a data object would point to a sequence group, and that
sequence group would point to the data object.
Reference counting doesn't save us in this situation. It
may be beneficial to preserve the overall logic for
simplicity, but right now everything related to this
relationship is purely internal and so a small rework was
possible to avoid cyclic references.

Closes #14.