Introduce a basic test system
This is a small bespoke test system based on
Godot's ability to run custom main loops from
CLI. It should be sufficient enough for basic
checks, and also should allow us some more
custom validation later on — to test the synth
itself for regressions.

This is just a basic implementation for now, and
as such it is not added to CI just yet.