Change semantics of the voice preset util
Instead of always initializing it with full set of
voices via the default constructor, which defeats
the purpose of using the flags, expose the
generator method as static and use it to create
an instance of the util.

This is a breaking change, but with minimal
impact. Simply changing .new() for
.generate_voices() should be sufficient for the
most basic case.