Rework the example project
Splits the example project into two distinct
modes: one with the piano roll and the other with
the MML player. We load a list of archived MML
songs composed for the original SiON for the
MML player.

Also changes the driver to clear the data object
when stop() is called, which is different from the
original SiON, but seems like an oversight. If we
don't do that, calling play() without any
parameters after calling it with an MML string
plays the compiled MML song over again. This
doesn't really make practical sense, as we need
to be able to actually stop the driver fully
somehow.

Note that some MML songs that we load aren't
valid (or there are bugs), so something needs to
be done about that.