Abstract access to the value in SLL
This is a part of a rework that should make it
safer to use and properly managed.

The final goal of this rework is to have one object
per list/ring, and keep actual data elements inside
of it. Right now there is no object representing the
list itself, as every SLL instance is a data object that
is linked to another SLL instance. This means that
nobody has the ownership over the list itself, which
makes it extremely easy to mismanage memory.

With a proper separation we should be able to
make it easier to maintain lists and clean them up
when we need to. As was the case with previous
commits, this change here aims to organize the
access to the internal parts of the list, so the future
rewrite is easier to perform.