Rewrite the singly-linked list implementation
This is the final step in the rework of the SLL type.
After a few preliminary changes to robustify the
API, I ended up doing a full rewrite of the type,
updating every class that uses it with the new
implementation in mind.

Because of how the SLL is used, some places are
still a bit manual and don't use some high-level
methods. This can probably be improved at some
point, but that's not needed right now.

Memory management has also been improved
while maintaining the shared pool to reduce the
number of allocations.

This implementation has been tested with the
example project and after a few fixes seems to be
working correctly. Tests also pass now, after the
same fixes. But I don't have full coverage to be
100% sure it's perfectly fine. So, probably expect
some bugs?

That said, during this rework I actually fixed a
couple of issues that remained unnoticed all this
time. So things should be even better than before
now.