Make stringification methods more Godot-y
Also slightly expose operators/operator
params. These classes had stringification, so
they had to be converted to Godot objects
(which should've been done all along anyway).
But no actual API has been exposed yet.

For the operator class I don't know if we want
to expose anything at all (neither ops nor
channels are actually accessible through
scripting at this time). But for the params we
should expose properties/methods the same
way we did for the channel params. It's actually
useful, but has to be done with documentation.
And I don't have time to do that right now.
So, for the later.