Enumerate and document all MA-3 wave shapes
This is a breaking change because of a rename.
PULSE_MA3_WAVE is now PULSE_MA3_SINE.
The rest of the values are new, although they
they only add definitions for values which
were allowed previously anyway. So that part
is not breaking.

Also some minor cleanup related to MA-3
and wave shapes.