Clean up some confusing identifiers around detune parameters
There was no consistency with what was called detune,
detune1, dt1, dt2. I believe the root of this issue comes
from the fact that we have a unified operator
implementation that is used by every different channel
implementation. As such, some terms become harder
to distinguish.

I tried to give more unique names where that made
sense, based on the existing code comments. And to
ensure some consistency with names like detune1 and
detune2.

These changes are internal, as none of the related
parameters have been exposed yet (partially, because
of this confusion being present).