Clean up some voice names in the voice preset util
This is mostly about normalization of the
names, fixing formatting, etc. For MIDI
drums I also aligned them with the Standard
MIDI spec, to the best of my knowledge.

A lot of voices have various hints and tuning
details encoded in their names. Perhaps this
should be properly extracted somewhere
instead of mixing with names?