Refactor SiONDriver.play into two separate method
This helps clean up the semantics of how
the driver works. We now have a dedicated
stream method, which can be called to just
start streaming without any data. At the
same time the old play method now expects
data for processing.

This is accompanied by two new methods,
is_streaming and is_paused, which report on
the corresponding statuses.

PS. This has a minor compatibility
implication. It is impossible now to call
SiONDriver.play without any arguments. But
the data argument itself is not enforced, and
can be null. Meaning that play(null) can still
be used to achieve the same effect as
stream().